% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.regression.tree.R
\name{evaluate_regression_tree}
\alias{evaluate_regression_tree}
\title{Evaluate a pre-specified regression tree}
\usage{
evaluate_regression_tree(
  LF,
  fcol,
  lcol,
  Flagcol,
  bins,
  save_dir,
  folder_name,
  pdf = FALSE,
  lf_scale = 1.5
)
}
\arguments{
\item{LF}{The length frequency data frame input; must include four columns: lat, lon, year, and quarter}

\item{fcol}{The first column in the data frame with length frequency info}

\item{lcol}{The last column in the data frame with length frequency info}

\item{Flagcol}{The column for pre-defined fishery definitions}

\item{bins}{Names of all bins included in LF}

\item{save_dir}{The directory where results will be saved}

\item{folder_name}{The name of the folder where results will be saved}

\item{pdf}{Whether to save figures in pdf - default is in png}

\item{lf_scale}{scale mean lf plot}
}
\description{
\code{evaluate.regression.tree} This function is the main regression tree function
}
