% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.klderror.null.R
\name{get.klderror.null}
\alias{get.klderror.null}
\title{Run the regression tree analysis with specified number of splits}
\usage{
get.klderror.null(response.mat, weight)
}
\description{
\code{get.klderror.null} This function generates an area code for longline catch allocation
}
