% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.kld.FINAL.R
\name{imp.kld.FINAL.R}
\alias{imp.kld.FINAL.R}
\title{Run the regression tree analysis with specified number of splits}
\usage{
imp.kld.FINAL.R(lf.y, lf.splitflg)
}
\description{
\code{imp.kld.FINAL.R} This function generates an area code for longline catch allocation
}
