% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.demean.R
\name{lf.demean}
\alias{lf.demean}
\title{Aggregate length counts data}
\usage{
lf.demean(LF, fcol, lcol, bins)
}
\arguments{
\item{LF}{The length frequency data frame input; must include four columns: lat, lon, year, and quarter}

\item{fcol}{The first column in the data frame with length frequency info}

\item{lcol}{The last column in the data frame with length frequency info}

\item{bins}{Names of all bins included in LF}
}
\description{
\code{lf.demean} aggregates counts data by year, quarter, lat, lon, and user-sepcified length bins
}
