% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.lf.map.R
\name{make.lf.map}
\alias{make.lf.map}
\title{Visualize the input LF data}
\usage{
make.lf.map(
  LF,
  fcol,
  lcol,
  bins,
  save_dir,
  plot_name = "LF_map",
  plot_format = "png",
  width = length(unique(LF$lon)),
  height = length(unique(LF$lat)),
  lwd = 1
)
}
\arguments{
\item{LF}{The length frequency data frame input; must include four columns: lat, lon, year, and quarter}

\item{fcol}{The first column in the data frame with length frequency info}

\item{lcol}{The last column in the data frame with length frequency info}

\item{bins}{Names of all bins included in LF}

\item{save_dir}{The directory where results will be saved}

\item{plot_name}{The name of the plot to be saved}

\item{plot_format}{The format (e.g., png and pdf) of the plot to be saved}

\item{width}{The width of the plot to be saved}

\item{height}{The height of the plot to be saved}

\item{lwd}{The width of lf bars}
}
\description{
\code{make.lf.map} This function plot the input LF data as maps
}
