% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.split.map.R
\name{make.split.map}
\alias{make.split.map}
\title{Map the split outcome}
\usage{
make.split.map(
  LF,
  Nsplit,
  save_dir,
  plot_name = "Split_map",
  plot_format = "png",
  width = length(unique(LF$lon))/2,
  height = length(unique(LF$lat))/2,
  s = 10
)
}
\arguments{
\item{LF}{The length frequency data frame input; must include four columns: lat, lon, year, and quarter}

\item{Nsplit}{The number of splits}

\item{save_dir}{The directory where results will be saved}

\item{plot_name}{The name of the plot to be saved}

\item{plot_format}{The format (e.g., png and pdf) of the plot to be saved}

\item{width}{The width of the plot to be saved}

\item{height}{The height of the plot to be saved}

\item{s}{The size of squares}
}
\description{
\code{make.split.map} This function color-codes cells based on the regression tree outcome
}
