% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.Us.areaflags.R
\name{make.Us.areaflags.f}
\alias{make.Us.areaflags.f}
\title{Run the regression tree analysis with specified number of splits}
\usage{
make.Us.areaflags.f(lf.input, key, value, split_num, area_num)
}
\description{
\code{make.Us.areaflags.f} This function generates an area code for longline catch allocation
}
