% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_CQrt.R
\name{rename_CQrt}
\alias{rename_CQrt}
\title{Run the regression tree analysis with specified number of splits}
\usage{
rename_CQrt(split)
}
\description{
\code{rename_CQrt} This function finds the best split based on the improvement
}
