% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.regression.tree.R
\name{run_regression_tree}
\alias{run_regression_tree}
\title{Run the regression tree analysis with specified number of splits}
\usage{
run_regression_tree(
  LF,
  fcol,
  lcol,
  Nsplit,
  save_dir,
  manual = FALSE,
  select = NA,
  lat.min = 1,
  lon.min = 1,
  quarter = TRUE
)
}
\arguments{
\item{LF}{The length frequency data frame input; must include columns lat, lon, year, and quarter}

\item{fcol}{The first column in the data frame with length frequency info}

\item{lcol}{The first column in the data frame with length frequency info}

\item{Nsplit}{The number of splits}

\item{save_dir}{The directory where results will be saved}

\item{manual}{Whether to use user-specified splits; default = FALSE}

\item{select}{User-specified splits; default = NA}

\item{quarter}{Whether to consider quarter as a splitting dimention; default = TRUE}
}
\value{
return the input LF with cell number and the percentage of total LF variance explained by each split
}
\description{
\code{run_regression_tree} This function generates an area code for longline catch allocation
}
