% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.tree.kld.FINAL.R
\name{simult.tree.kld.FINAL}
\alias{simult.tree.kld.FINAL}
\title{Run the regression tree analysis with specified number of splits}
\usage{
simult.tree.kld.FINAL(
  lfinput.frm,
  frstcol.lf,
  lstcol.lf,
  lat.min = lat.min,
  lon.min = lon.min
)
}
\arguments{
\item{lfinput.frm}{The length frequency data frame input; must include columns lat, lon, year, and quarter}

\item{frstcol.lf}{The first column in the data frame with length frequency info}

\item{lstcol.lf}{The first column in the data frame with length frequency info}
}
\description{
\code{simult.tree.kld.FINAL} This function finds the best split based on the improvement
}
