% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_split.R
\name{find_split}
\alias{find_split}
\title{Run the regression tree analysis with specified number of splits}
\usage{
find_split(lf_prop, frstcol.lf, lstcol.lf, lat.min, lon.min, quarter)
}
\arguments{
\item{lf_prop}{The length frequency data frame input; must include columns lat, lon, year, and quarter}

\item{frstcol.lf}{The first column in the data frame with length frequency info}

\item{lstcol.lf}{The first column in the data frame with length frequency info}

\item{quarter}{Whether to consider quarter as a splitting dimention; default = TRUE}
}
\description{
\code{find_split} This function finds the best split based on the improvement
}
