% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.aggregate.R
\name{lf.aggregate}
\alias{lf.aggregate}
\title{Aggregate length counts data}
\usage{
lf.aggregate(LC, fcol, lcol, bins, new_bins, LengthOnly = FALSE, minN = 0)
}
\arguments{
\item{LC}{The length frequency data frame input; must include four columns: lat, lon, year, and quarter}

\item{fcol}{The first column in the data frame with length frequency info}

\item{lcol}{The last column in the data frame with length frequency info}

\item{bins}{Names of all bins included in LF}

\item{new_bins}{Names of all bins included in LF}

\item{LengthOnly}{Whether aggregate data by new length bins only}

\item{minN}{minimal sample size required for a year-quarter-lat-lon stratum}
}
\description{
\code{lf.aggregate} aggregates counts data by year, quarter, lat, lon, and user-sepcified length bins
}
