% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.regression.tree.R
\name{loop_regression_tree}
\alias{loop_regression_tree}
\title{Loop the regression tree analysis with specified number of splits}
\usage{
loop_regression_tree(
  LF,
  fcol,
  lcol,
  Nsplit,
  save_dir,
  max_select,
  lat.min = 1,
  lon.min = 1,
  quarter = TRUE,
  include_dummy = FALSE
)
}
\arguments{
\item{LF}{The length frequency data frame input; must include columns lat, lon, year, and quarter}

\item{fcol}{The first column in the data frame with length frequency info}

\item{lcol}{The first column in the data frame with length frequency info}

\item{Nsplit}{The number of splits}

\item{save_dir}{The directory where results will be saved}

\item{max_select}{User-specified number of splits to be explored for each split; for example, 2 means exploring the best 2 splits for every split}

\item{quarter}{Whether to consider quarter as a splitting dimention; default = TRUE}
}
\value{
return the input LF with cell number and the percentage of total LF variance explained by each split
}
\description{
\code{loop_regression_tree} This function generates an area code for longline catch allocation
}
