% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.regression.tree.R
\name{run_regression_tree}
\alias{run_regression_tree}
\title{Run the regression tree analysis with specified number of splits}
\usage{
run_regression_tree(
  LF,
  fcol,
  lcol,
  bins,
  Nsplit,
  save_dir,
  manual = FALSE,
  select = NA,
  lat.min = 1,
  lon.min = 1,
  year.min = 1,
  quarter = TRUE,
  year = FALSE,
  include_dummy = FALSE,
  pdf = FALSE,
  lf_scale = 2
)
}
\arguments{
\item{LF}{The length frequency data frame input; must include four columns: lat, lon, year, and quarter}

\item{fcol}{The first column in the data frame with length frequency info}

\item{lcol}{The last column in the data frame with length frequency info}

\item{bins}{Names of all bins included in LF}

\item{Nsplit}{The number of splits}

\item{save_dir}{The directory where results will be saved}

\item{manual}{Whether to use user-specified splits; default = FALSE}

\item{select}{User-specified splits; default = NA}

\item{lat.min}{Minimal number of lat grids allowed for a cell, which could not be split beyond that}

\item{lon.min}{Minimal number of lon grids allowed for a cell, which could not be split beyond that}

\item{year.min}{Minimal number of years allowed for a cell, which could not be split beyond that}

\item{quarter}{Whether to consider quarter as a splitting dimension; default = TRUE}

\item{year}{Whether to consider year as a splitting dimension; default = FALSE}

\item{include_dummy}{Whether to include dummy data; default = FALSE}

\item{pdf}{Whether to save figures in pdf - default is in png}
}
\value{
return the input LF with cell number and the percentage of total LF variance explained by each split
}
\description{
\code{run.regression.tree} This function is the main regression tree function
}
